basic setup for google analytics

+++ app/src/main/java/com/blongdev/sift/SubredditInfo.java
@@ -5,7 +5,7 @@ package com.blongdev.sift;
  */
 public class SubredditInfo {
     public String mName;
-    public int mId;
+    public long mId;
     public int mPoints;
     public int mAge;
     public String mServerId;
